타인의 보유주식 리스트 보기 완료

+++ community-module/src/main/java/com/pda/community_module/domain/OwnStock.java
@@ -32,5 +32,5 @@ public class OwnStock extends BaseEntity {
 
     private Long stockCost;
 
-    private Long profit;
+    private Double profit;
 }
